Refactored AuthManager
