+++ CafeTime/CafeTime/UIViewController+Extension.swift
@@ -17,4 +17,8 @@ extension UIViewController {
         present(alertController, animated: true, completion: nil)
     }
     
+    func presentInNav(vcs: [UIViewController]) {
+        let nav = UINavigationController(rootViewController: vcs[1])
+        vcs[0].present(nav, animated: true, completion: nil)
+    }
 }
